+++ tests/formats/amber-netcdf.cpp
@@ -21,24 +21,9 @@ bool roughly(const Vector3D& a, const Vector3D& b, const double eps){
 
 TEST_CASE("Read files in NetCDF format", "[Amber NetCDF]"){
     Trajectory file(NCDIR "water.nc");
-    Frame frame;
 
-    SECTION("Stream style reading"){
-        file >> frame;
-        CHECK(frame.natoms() == 297);
-        // Check positions
-        auto positions = frame.positions();
-
-        CHECK(roughly(positions[0], vector3d(0.4172191f, 8.303366f, 11.73717f), 1e-4));
-        CHECK(roughly(positions[296], vector3d(6.664049f, 11.61418f, 12.96149f), 1e-4));
-
-        auto cell = frame.cell();
-        CHECK(cell.type() == UnitCell::ORTHOROMBIC);
-        CHECK(fabs(cell.a() - 15.0) < EPS);
-    }
-
-    SECTION("Method style reading"){
-        frame = file.read();
+    SECTION("Read one frame"){
+        auto frame = file.read();
         CHECK(frame.natoms() == 297);
         // Check positions
         auto positions = frame.positions();
@@ -47,17 +32,17 @@ TEST_CASE("Read files in NetCDF format", "[Amber NetCDF]"){
     }
 
     SECTION("Read more than one frame"){
-        file >> frame;
-        file >> frame;
-        file >> frame;
+        auto frame = file.read();
+        frame = file.read();
+        frame = file.read();
         CHECK(frame.natoms() == 297);
 
         auto positions = frame.positions();
         CHECK(roughly(positions[0], vector3d(0.2990952f, 8.31003f, 11.72146f), 1e-4));
         CHECK(roughly(positions[296], vector3d(6.797599f, 11.50882f, 12.70423f), 1e-4));
 
-        while (not file.done()){
-            file >> frame;
+        while (!file.done()){
+            frame = file.read();
         }
         positions = frame.positions();
         CHECK(roughly(positions[0], vector3d(0.3185586f, 8.776042f, 11.8927f), 1e-4));
@@ -75,7 +60,7 @@ TEST_CASE("Write files in NetCDF format", "[Amber NetCDF]"){
             positions[i] = vector3d(1, 2, 3);
 
         frame.set_topology(dummy_topology(4));
-        file << frame;
+        file.write(frame);
     }
 
     SECTION("Check the file") {